move loadArea to /src/helpers
